+++ model/schema_order_test.go
@@ -11,4 +11,5 @@ func TestGetOrder(t *testing.T) {
 func TestTopOrder(t *testing.T) {
 	// TODO: Schema.TopAscRows, TopAscRowsAt
 	// TODO: Schema.TopDescRows, TopDescRowsAt
+	// TODO: Schema.TopOrderedRows, Schema.TopOrderedRowsAt
 }