+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.3.31 - Multi-Column Order
-  * Commit:
+  * Commit: 2026-05-18 09:56
   * Schema GetOrderedRows
+  * Schema TopOrderedRows
 ## v0.3.30 - Secrets Package 
   * Commit: 2026-05-16 23:38
   * Secrets package